bigint -> tinyint

+++ src/main/resources/sql/maxwell_schema.sql
@@ -40,7 +40,7 @@ CREATE TABLE IF NOT EXISTS `columns` (
   coltype     varchar(255),
   is_signed   tinyint(1) unsigned,
   enum_values text,
-  column_length bigint unsigned,
+  column_length tinyint unsigned,
   index (schema_id),
   index (table_id)
 );
